Source code of Bottema in Maple - Prof. Lu Yang
Automated Proving for a Class of Constructive Geometric Inequalities. An automated inequality-proving algorithm is presented based on a mixed method including a so-called cell-decomposition. That is implemented by a Maple program named “BOTTEMA” which can prove or disprove propositions in an extensive class of geometric and algebraic inequalities involving radicals. Most of the theorems in “Geometric Inequalities”　writed by Bottema et al., can be proven efficiently in this way.